// frontend/plugins/welcome/src/api/models/EntEmployee.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SUT SA Example API Patient
 * This is a sample server for SUT SE 2563
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    EntEmployeeEdges,
    EntEmployeeEdgesFromJSON,
    EntEmployeeEdgesFromJSONTyped,
    EntEmployeeEdgesToJSON,
} from './';

/**
 * 
 * @export
 * @interface EntEmployee
 */
export interface EntEmployee {
    /**
     * BirthdayDate holds the value of the "BirthdayDate" field.
     * @type {string}
     * @memberof EntEmployee
     */
    birthdayDate?: string;
    /**
     * Email holds the value of the "Email" field.
     * @type {string}
     * @memberof EntEmployee
     */
    email?: string;
    /**
     * EmployeeName holds the value of the "EmployeeName" field.
     * @type {string}
     * @memberof EntEmployee
     */
    employeeName?: string;
    /**
     * Password holds the value of the "Password" field.
     * @type {string}
     * @memberof EntEmployee
     */
    password?: string;
    /**
     * Tel holds the value of the "Tel" field.
     * @type {string}
     * @memberof EntEmployee
     */
    tel?: string;
    /**
     * UserId holds the value of the "UserId" field.
     * @type {string}
     * @memberof EntEmployee
     */
    userId?: string;
    /**
     * 
     * @type {EntEmployeeEdges}
     * @memberof EntEmployee
     */
    edges?: EntEmployeeEdges;
    /**
     * ID of the ent.
     * @type {number}
     * @memberof EntEmployee
     */
    id?: number;
}

export function EntEmployeeFromJSON(json: any): EntEmployee {
    return EntEmployeeFromJSONTyped(json, false);
}

export function EntEmployeeFromJSONTyped(json: any, ignoreDiscriminator: boolean): EntEmployee {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'birthdayDate': !exists(json, 'BirthdayDate') ? undefined : json['BirthdayDate'],
        'email': !exists(json, 'Email') ? undefined : json['Email'],
        'employeeName': !exists(json, 'EmployeeName') ? undefined : json['EmployeeName'],
        'password': !exists(json, 'Password') ? undefined : json['Password'],
        'tel': !exists(json, 'Tel') ? undefined : json['Tel'],
        'userId': !exists(json, 'UserId') ? undefined : json['UserId'],
        'edges': !exists(json, 'edges') ? undefined : EntEmployeeEdgesFromJSON(json['edges']),
        'id': !exists(json, 'id') ? undefined : json['id'],
    };
}

export function EntEmployeeToJSON(value?: EntEmployee | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'BirthdayDate': value.birthdayDate,
        'Email': value.email,
        'EmployeeName': value.employeeName,
        'Password': value.password,
        'Tel': value.tel,
        'UserId': value.userId,
        'edges': EntEmployeeEdgesToJSON(value.edges),
        'id': value.id,
    };
}