// frontend/plugins/welcome/src/api/models/EntNametitle.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SUT SA Example API Patient
 * This is a sample server for SUT SE 2563
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    EntNametitleEdges,
    EntNametitleEdgesFromJSON,
    EntNametitleEdgesFromJSONTyped,
    EntNametitleEdgesToJSON,
} from './';

/**
 * 
 * @export
 * @interface EntNametitle
 */
export interface EntNametitle {
    /**
     * Title holds the value of the "Title" field.
     * @type {string}
     * @memberof EntNametitle
     */
    title?: string;
    /**
     * 
     * @type {EntNametitleEdges}
     * @memberof EntNametitle
     */
    edges?: EntNametitleEdges;
    /**
     * ID of the ent.
     * @type {number}
     * @memberof EntNametitle
     */
    id?: number;
}

export function EntNametitleFromJSON(json: any): EntNametitle {
    return EntNametitleFromJSONTyped(json, false);
}

export function EntNametitleFromJSONTyped(json: any, ignoreDiscriminator: boolean): EntNametitle {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'title': !exists(json, 'Title') ? undefined : json['Title'],
        'edges': !exists(json, 'edges') ? undefined : EntNametitleEdgesFromJSON(json['edges']),
        'id': !exists(json, 'id') ? undefined : json['id'],
    };
}

export function EntNametitleToJSON(value?: EntNametitle | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'Title': value.title,
        'edges': EntNametitleEdgesToJSON(value.edges),
        'id': value.id,
    };
}